# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval — Tidebreak API
 */
export interface HealthStatus {
  status: string;
}

export interface AuthCredentials {
  /** @minLength 3 @maxLength 32 */
  username: string;
  /** @minLength 6 */
  password: string;
}

export interface AuthUser {
  id: number;
  username: string;
  createdAt: string;
}

export interface SaveData {
  userId: number;
  highScore: number;
  pearls: number;
  upgradesJson: string;
  currentWave: number;
  updatedAt: string;
}

export interface SaveInput {
  highScore?: number;
  pearls?: number;
  upgradesJson?: string;
  currentWave?: number;
}

export interface LeaderboardEntry {
  rank: number;
  username: string;
  highScore: number;
  currentWave: number;
}

export interface MessageResponse {
  message: string;
}

export interface ErrorResponse {
  error: string;
}