# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: INVASOL GAMES — Tidebreak API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: save
    description: Save data operations
  - name: leaderboard
    description: Leaderboard operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthCredentials"
      responses:
        "201":
          description: Registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Username already taken
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with credentials
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthCredentials"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout current session
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current logged-in user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /save:
    get:
      operationId: getSave
      tags: [save]
      summary: Get save data for the logged-in user
      responses:
        "200":
          description: Save data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaveData"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: upsertSave
      tags: [save]
      summary: Upsert save data for the logged-in user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveInput"
      responses:
        "200":
          description: Saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaveData"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /leaderboard:
    get:
      operationId: getLeaderboard
      tags: [leaderboard]
      summary: Get top high scores
      responses:
        "200":
          description: Top scores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaderboardEntry"

components:
  schemas:
    HealthStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string

    AuthCredentials:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 32
        password:
          type: string
          minLength: 6

    AuthUser:
      type: object
      required: [id, username, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        createdAt:
          type: string

    SaveData:
      type: object
      required: [userId, highScore, pearls, upgradesJson, currentWave, updatedAt]
      properties:
        userId:
          type: integer
        highScore:
          type: integer
        pearls:
          type: integer
        upgradesJson:
          type: string
        currentWave:
          type: integer
        updatedAt:
          type: string

    SaveInput:
      type: object
      properties:
        highScore:
          type: integer
        pearls:
          type: integer
        upgradesJson:
          type: string
        currentWave:
          type: integer

    LeaderboardEntry:
      type: object
      required: [rank, username, highScore, currentWave]
      properties:
        rank:
          type: integer
        username:
          type: string
        highScore:
          type: integer
        currentWave:
          type: integer

    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string